Merge PR https://github.com/OCA/credit-control/pull/86 into 12.0-prg_master
[IMP+FIX] account_credit_control: traceable communications

Before this patch, `credit.control.communication` was a `TransientModel` that sent emails. This produced that, when someone responded to one of those emails, Odoo detected it was a response but couldn't know where to store it, so it was lost. Also, users with credit control access got a 500 error when clicking on the "View" button in the email.

Now, `credit.control.communication` is a normal `Model` which inherits from `mail.thread` and `mail.activity.mixin`. This makes more sense, because when you are tracking credit control communications, you usually communicate with a customer regarding several lines at once (no sense on asking a customer to pay each line separately).

More changes:

- `credit.control.run` now displays an additional smart button that leads users to generated communications.
- `credit.control.line` now is linked to a communication instead of a simple message (the old link is preserved in a hidden legacy column).
- `credit.control.line` has a new state "queued", which indicates that the communication is executed but still not sent. When sent, it's changed to "done" or "email_error".
- Comunication emails have a specific subtype.
- Some fields are renamed, as they didn't follow guidelines.
- Some menu names are less verbose now.
- Mail templates code is migrated.

@Tecnativa TT24841