[18.0][FIX] account_financial_risk: Fix access rights on credit_limit
When users without accounting access are trying to access a Sale Order, they have an access error on the partner field credit_limit (standard Odoo field that is limited to Account groups), which is used when computing the SO's risk_info. This is fixed by adding the group financial_risk_user to the field. Now the user can access the SO, and sees the risk_info if they have the financial_risk_user right, otherwise it is not displayed, with no access error.